move models into dedicated namespace + use indentifier VOs + add a StandardUser
