--- src/Domain/Organization.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Shippeo\Heimdall\Domain;
-
-interface Organization
-{
-    public function id(): string;
-}
